cleaned up the oauth mapper-madness - ToDo: add error case in oauth.service requestToken() test
