+++ apps/server/src/modules/oauth/controller/dto/oauth.dto.spec.ts
@@ -1,29 +1,18 @@
 import { validate } from 'class-validator';
-import { TokenRequestParams } from './token-request.params';
 import { TokenRequestPayload } from './token-request.payload';
 import { OauthTokenResponse } from './oauth-token.response';
 
-const tokenRequestParams = new TokenRequestParams();
-tokenRequestParams.client_id = '12345';
-tokenRequestParams.client_secret = 'asdf';
-tokenRequestParams.code = '1111';
-tokenRequestParams.grant_type = 'code';
-tokenRequestParams.redirect_uri = 'www.mock.de';
-
-describe('token-request-params', () => {
-	it('should validate', async () => {
-		const validationErrors = await validate(tokenRequestParams);
-		expect(validationErrors).toHaveLength(0);
-	});
-});
+const tokenRequestPayload = new TokenRequestPayload();
+tokenRequestPayload.client_id = '12345';
+tokenRequestPayload.client_secret = 'asdf';
+tokenRequestPayload.code = '1111';
+tokenRequestPayload.grant_type = 'code';
+tokenRequestPayload.redirect_uri = 'www.mock.de';
+tokenRequestPayload.tokenEndpoint = 'www.tokenendpointmock,de';
 
 describe('token-request-payload', () => {
-	const tokenRequestPayload = new TokenRequestPayload();
-	tokenRequestPayload.tokenEndpoint = 'asdfgh';
-	tokenRequestPayload.tokenRequestParams = tokenRequestParams;
-
 	it('should validate', async () => {
-		const validationErrors = await validate(tokenRequestParams);
+		const validationErrors = await validate(tokenRequestPayload);
 		expect(validationErrors).toHaveLength(0);
 	});
 });
