+++ apps/server/src/modules/oauth/service/oauth.service.spec.ts
@@ -194,7 +194,6 @@ describe('OAuthService', () => {
 				clientSecret: 'mocksecret',
 				tokenEndpoint: 'http://mock.de/mock/auth/public/mockToken',
 				grantType: 'authorization_code',
-				tokenRedirectUri: 'http://mockhost:3030/api/v3/oauth/testsystemId/token',
 				scope: 'openid uuid',
 				responseType: 'code',
 				authEndpoint: 'mock_authEndpoint',
@@ -202,7 +201,7 @@ describe('OAuthService', () => {
 				logoutEndpoint: 'logoutEndpointMock',
 				issuer: 'mock_issuer',
 				jwksEndpoint: 'mock_jwksEndpoint',
-				codeRedirectUri: 'http://mockhost:3030/api/v3/oauth/testsystemId',
+				redirectUri: 'http://mockhost:3030/api/v3/oauth/testsystemId',
 			},
 			_id: new ObjectId(),
 			id: '',
