+++ apps/server/src/modules/system/mapper/system.mapper.spec.ts
@@ -61,7 +61,6 @@ describe('SystemMapper', () => {
 			// Assert
 			expect(result[0].oauthConfig?.clientId).toEqual(systemEntities[0].oauthConfig?.clientId);
 			expect(result[0].oauthConfig?.clientSecret).toEqual(systemEntities[0].oauthConfig?.clientSecret);
-			expect(result[0].oauthConfig?.tokenRedirectUri).toEqual(systemEntities[0].oauthConfig?.tokenRedirectUri);
 			expect(result[0].oauthConfig?.grantType).toEqual(systemEntities[0].oauthConfig?.grantType);
 			expect(result[0].oauthConfig?.tokenEndpoint).toEqual(systemEntities[0].oauthConfig?.tokenEndpoint);
 			expect(result[0].oauthConfig?.authEndpoint).toEqual(systemEntities[0].oauthConfig?.authEndpoint);
@@ -71,7 +70,7 @@ describe('SystemMapper', () => {
 			expect(result[0].oauthConfig?.logoutEndpoint).toEqual(systemEntities[0].oauthConfig?.logoutEndpoint);
 			expect(result[0].oauthConfig?.issuer).toEqual(systemEntities[0].oauthConfig?.issuer);
 			expect(result[0].oauthConfig?.jwksEndpoint).toEqual(systemEntities[0].oauthConfig?.jwksEndpoint);
-			expect(result[0].oauthConfig?.codeRedirectUri).toEqual(systemEntities[0].oauthConfig?.codeRedirectUri);
+			expect(result[0].oauthConfig?.redirectUri).toEqual(systemEntities[0].oauthConfig?.redirectUri);
 			expect(result[1].oauthConfig).toBe(undefined);
 		});
 	});
